smartcontract

+++ contracts/Token.sol
@@ -22,7 +22,7 @@ contract Token {
         balances[to] += amount;
     }
 
-    function getBalance(address account) external view returns(uint) {
+    function getBalance(address account) external view returns(uint256) {
         return balances[account];
     }
